separated all database scrapers to their own files, only import their functionality if needed at runtime.
